requirements.txt
python-dotenv
# openai
google-generativeai
pyupbit
pyjwt
pandas
pandas_ta
schedule
streamlit
selenium